Solution to Check Leap Year

+++ Check Leap Year@@6788dd63490637cfc44e1b90/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a;
-    scanf("%d",&a);
+    double a;
+    scanf("%lf",&a);
     if(a%4==0 && a%400==0){
         printf(Leap Year);
     }else{
